feat(Auth HOC): #Y-31 Авторизация должна проверяться через отдельно написанный Hook или HOC
